requirements.txt
tensorflow
numpy
pandas
scikit-learn
matplotlib
opencv-python
Pillow
streamlit
plotly